refactor(internal/sample): merge sample packages
Merge internal/sidekick/sample to internal/sample so that there is only
one package for sample testdata.
