+++ internal/sidekick/parser/svcconfig/svcconfig_test.go
@@ -18,8 +18,8 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
+	"github.com/googleapis/librarian/internal/sample"
 	"github.com/googleapis/librarian/internal/serviceconfig"
-	"github.com/googleapis/librarian/internal/sidekick/sample"
 )
 
 func TestExtractPackageName(t *testing.T) {
